requirements.txt
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
executing==2.0.1
filelock==3.13.3
fonttools==4.47.2
idna==3.6
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
nest-asyncio==1.5.9
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pymemcache==4.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-memcached==1.62
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0n
SCons==4.6.0.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
